Move DaysOfWeek model to core app
--HG--
rename : wger/manager/fixtures/days_of_week.json => wger/core/fixtures/days_of_week.json
extra : amend_source : ee928aa596247c5282620116529a85d7ab3552df
