+++ wger/manager/fixtures/test-user-data.json
@@ -128,17 +128,17 @@
                 ], 
                 [
                     "add_daysofweek", 
-                    "manager", 
+                    "core",
                     "daysofweek"
                 ], 
                 [
                     "change_daysofweek", 
-                    "manager", 
+                    "core",
                     "daysofweek"
                 ], 
                 [
                     "delete_daysofweek", 
-                    "manager", 
+                    "core",
                     "daysofweek"
                 ], 
                 [
