+++ wger/urls.py
@@ -28,7 +28,7 @@ from wger.nutrition.api import resources as nutrition_api
 from wger.utils.generic_views import TextTemplateView
 from wger.utils.generic_views import WebappManifestView
 from wger.manager.api import resources as manager_api
-from wger.core.api.resources import UserProfileResource, LanguageResource
+from wger.core.api.resources import UserProfileResource, LanguageResource, DaysOfWeekResource
 from wger.weight.api import resources as weight_api
 
 #
@@ -58,7 +58,7 @@ v1_api.register(UserProfileResource())
 v1_api.register(manager_api.WorkoutResource())
 v1_api.register(manager_api.ScheduleResource())
 v1_api.register(manager_api.ScheduleStepResource())
-v1_api.register(manager_api.DaysOfWeekResource())
+v1_api.register(DaysOfWeekResource())
 v1_api.register(manager_api.DayResource())
 v1_api.register(manager_api.SetResource())
 v1_api.register(manager_api.SettingResource())
